refactor(internal-cloud-client): isolate ChannelInfo schema
Move GenerationInfo into ChannelInfo as inner class to make clarify
intent of where this schema class should be used.
