+++ internal-cloud-client/src/main/java/io/zeebe/clustertestbench/internal/cloud/InternalCloudAPIClient.java
@@ -3,7 +3,7 @@ package io.zeebe.clustertestbench.internal.cloud;
 import io.zeebe.clustertestbench.internal.cloud.request.CreateGenerationRequest;
 import io.zeebe.clustertestbench.internal.cloud.request.UpdateChannelRequest;
 import io.zeebe.clustertestbench.internal.cloud.response.ChannelInfo;
-import io.zeebe.clustertestbench.internal.cloud.response.GenerationInfo;
+import io.zeebe.clustertestbench.internal.cloud.response.ChannelInfo.GenerationInfo;
 import java.util.List;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
