+++ internal-cloud-client/src/test/java/io/zeebe/clustertestbench/internal/cloud/StubInternalCloudAPIClient.java
@@ -3,7 +3,7 @@ package io.zeebe.clustertestbench.internal.cloud;
 import io.zeebe.clustertestbench.internal.cloud.request.CreateGenerationRequest;
 import io.zeebe.clustertestbench.internal.cloud.request.UpdateChannelRequest;
 import io.zeebe.clustertestbench.internal.cloud.response.ChannelInfo;
-import io.zeebe.clustertestbench.internal.cloud.response.GenerationInfo;
+import io.zeebe.clustertestbench.internal.cloud.response.ChannelInfo.GenerationInfo;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
